backend/requirements/requirements.txt
fastapi
psycopg[pool]
uvicorn[standard]
python-jose
python-multipart
bcrypt
